% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetQuery_levelplots.R
\name{dbGetQuery_levelplots}
\alias{dbGetQuery_levelplots}
\title{dbGetQuery levelplots}
\usage{
dbGetQuery_levelplots(avoidpseudorep = TRUE,
  PARstarttime = "2018-10-10 15:00:00",
  PARendtime = "2018-10-11 09:00:00", PARaltbin = 50,
  PARtimebin = 15, PARaltmax = 500, PARclassification_id = "5,8",
  PARdegrees = c(0, 135, -54, -45))
}
\arguments{
\item{PARstarttime}{Start time of the radar data that should be extracted}

\item{PARendtime}{End time of the radar data that should be extracted}

\item{PARaltbin}{The aggregation bins for altitude in meters (50 meter)}

\item{PARtimebin}{The aggregation bins for the time in minutes (15 minutes)}

\item{PARaltmax}{The maximum altitude in meters for which metrics are extracted (500 meter)}

\item{PARclassification_id}{Classification_ids of interest (see database for the meaning of the codes)}

\item{PARdegrees}{The angles of interest in degrees}

\item{pseudorep}{Account for pseudoreplication by averaging per track or not}
}
\description{
Export aggregated data from robin radar eemshaven database to generate level plots
}
\examples{
# connect to database 
library(RPostgreSQL)
con <- dbConnect('PostgreSQL', dbname='staticeemshaven', host='hostname', user='username', password = 'password')
#run functions with standard settings 
a <- dbGetQuery_levelplots()
get data 
data <- a$data
get SQL QUERY()
query <- a$query
}
\keyword{RpostgreSQL,}
\keyword{aggregate}
\keyword{radar,}
