% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelplots.R
\name{levelplots}
\alias{levelplots}
\title{levelplots}
\usage{
levelplots(data = density, autoscale = TRUE, main = "",
  col = colorRampPalette(c("yellow", "red", "black"))(50), at = seq(0,
  1000, 50))
}
\arguments{
\item{data}{A data frame obtained from reshape_df}

\item{autoscale}{if autoscale is true than the color scale bins and min-max is automatically set}

\item{main}{title of the plot}

\item{col}{colors to use}

\item{at}{the color bins}
}
\description{
Generate a levelplot
}
\examples{
connect to database 
library(RPostgreSQL)
con <- dbConnect('PostgreSQL', 
                 dbname='staticeemshaven', 
                 host='hostname', 
                 user='username', 
                 password = 'password')
df <- dbGetQuery_levelplots()
density_log <- reshape_df(df,
              rowvar='altitude_bin',
              colvar='timestamp_bin',
              variable='density_log')
direction <- reshape_df(df,
              rowvar='altitude_bin',
              colvar='timestamp_bin',
              variable='density')
plot_density_log <- levelplots(density_log,
                               autoscale=TRUE, 
                               main='density [log]', 
                               col= colorRampPalette(c('yellow','red','black'))(50))
plot_direction <- levelplots(data=direction,
                             autoscale=FALSE, 
                             main='direction [0=North]',
                             col= colorRampPalette(c('red', 'yellow', 'green', 'cyan', 'blue', 'pink', 'red'))(360),
                             at=seq(0, 360, length.out=360))
grid.arrange(plot_density_log, plot_direction)
}
