% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_df.R
\name{reshape_df}
\alias{reshape_df}
\title{reshape dataframe}
\usage{
reshape_df(data = df, rowvar = "altitude_bin",
  colvar = "timestamp_bin", variable = "speed")
}
\arguments{
\item{data}{A data frame obtained from dbGetQuery_levelplots}

\item{rowvar}{The variable that represents the rows in the reshaped data frame}

\item{colvar}{The variable that represents the columns the reshaped data frame}

\item{variable}{The variable representing the values in the reshaped data frame}
}
\description{
Uses a row and column and a parameter variable to reshape a data frame
}
\examples{
connect to database 
library(RPostgreSQL)
con <- dbConnect('PostgreSQL', 
                 dbname='staticeemshaven', 
                 host='hostname', 
                 user='username', 
                 password = 'password')
df <- dbGetQuery_levelplots()
density_log <- reshape_df(df,
              rowvar='altitude_bin',
              colvar='timestamp_bin',
              variable='density_log')
}
\keyword{aggregate,}
\keyword{radar,}
\keyword{reshape}
